[FIX] 修复sonarlint报错

+++ src/main/java/io/choerodon/test/manager/app/eventhandler/TestManagerEventHandler.java
@@ -130,7 +130,7 @@ public class TestManagerEventHandler {
 
         } catch (Exception e) {
             if (map != null) {
-                testPlanServcie.setPlanInitStatusFail(modelMapper.map(testPlanMapper.selectByPrimaryKey(map.get("new").longValue()), TestPlanVO.class));
+                testPlanServcie.setPlanInitStatusFail(modelMapper.map(testPlanMapper.selectByPrimaryKey(map.get("new")), TestPlanVO.class));
             }
             throw e;
         }
